Refactor plugin structure to make app easier to test
