+++ emergence_game/src/main.rs
@@ -1,5 +1,4 @@
 use bevy::prelude::*;
-use emergence_lib::*;
 
 fn main() {
     App::new()
@@ -8,12 +7,8 @@ fn main() {
             ..Default::default()
         })
         .add_plugins(DefaultPlugins)
-        .add_plugin(camera::CameraPlugin)
-        .add_plugin(cursor::CursorTilePosPlugin)
-        .add_plugin(hive_mind::HiveMindPlugin)
-        .add_plugin(terrain::generation::GenerationPlugin)
-        .add_plugin(organisms::structures::StructuresPlugin)
-        .add_plugin(organisms::units::UnitsPlugin)
-        .add_plugin(signals::SignalsPlugin)
+        .add_plugin(emergence_lib::GraphicsPlugin)
+        .add_plugin(emergence_lib::SimulationPlugin)
+        .add_plugin(emergence_lib::InteractionPlugin)
         .run();
 }
