+++ emergence_lib/Cargo.toml
@@ -17,3 +17,7 @@ indexmap = "1.9"
 dashmap = "5.4"
 leafwing-input-manager = "0.6.1"
 emergence_macros = { path = "../emergence_macros" }
+
+[dev-dependencies]
+# We need headless operation in tests
+bevy_ecs_tilemap = { git = "https://github.com/StarArawn/bevy_ecs_tilemap", branch = "main", default-features = false}
